Update MOST RECENT
When checking if the tag is a closed tag, it correctly recognizes which tags are closed. But for some reason, when I check the stack for the open tag, I don't think it recognizes the open tag, so it automatically goes to the "else" and prints that there is no match every time. And at the end when I print an error if there are any open tags left in the stack without a match, all of them print.